+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 - Move `FragmentKey` into `android` package.
 - Move BaseFormulaFragment into `android` package.
 - Move `DisposableScope` into `android` package.
+- Move `FormulaFragment` into `android` package.
+- Move `BackCallback` into `android` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.
