+++ formula-android-tests/src/test/java/com/instacart/formula/FragmentFlowRenderViewTest.kt
@@ -9,7 +9,7 @@ import com.google.common.truth.Truth.assertThat
 import com.instacart.formula.fragment.FragmentContract
 import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.android.FragmentKey
-import com.instacart.formula.integration.BackCallback
+import com.instacart.formula.android.BackCallback
 import com.instacart.formula.test.TestContract
 import com.instacart.formula.test.TestContractWithId
 import com.instacart.formula.test.TestFragmentActivity
