+++ formula-android/src/main/java/com/instacart/formula/android/FeatureFactory.kt
@@ -1,6 +1,5 @@
 package com.instacart.formula.android
 
-import com.instacart.formula.fragment.FormulaFragment
 import com.instacart.formula.fragment.FragmentFlowStore
 
 /**
